rename custom->lit
