+++ src/scene/shader-lib/programs/lit-options-utils.js
@@ -1,0 +1,33 @@
+import { LIGHTTYPE_DIRECTIONAL } from "../../constants.js";
+
+const LitOptionsUtils = {
+
+    // generate a key for the lit options
+    generateKey(options) {
+        return "lit" + Object.keys(options)
+            .sort()
+            .map((key) => {
+                if (key === "chunks") {
+                    return LitOptionsUtils.generateChunksKey(options);
+                } else if (key === "lights") {
+                    return LitOptionsUtils.generateLightsKey(options);
+                } else {
+                    return key + options[key];
+                }
+            })
+            .join("");
+    },
+
+    generateLightsKey(options) {
+        return options.lights.map(light => !options.clusteredLightingEnabled || light._type === LIGHTTYPE_DIRECTIONAL ? light.key : "").join("");
+    },
+
+    generateChunksKey(options) {
+        return Object.keys(options.chunks ?? {})
+            .sort()
+            .map(key => key + options.chunks[key])
+            .join("");
+    },
+};
+
+export { LitOptionsUtils };
